Add vcs-clear-retry option to reset the repositories
It avoids conflicts if a branch has diverges for instance after a rebase of a
merged pull request.
Only set on prod and test environments because on the dev one it could lead to
an incontrollable chain of consequences involving a mad developer following
several hours of painstakingly hand-crafted code lost with a single
bin/buildout.